Add custom hook for competitions
